moved all reservation queries to BookingSystem class and test files to booking_system_spec - tests pass
