+++ spec/spec_helper.rb
@@ -16,3 +16,4 @@ Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 # Require_relative your lib files here!
 require_relative '../lib/room'
 require_relative '../lib/reservation'
+require_relative '../lib/booking_system'
